requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
PyMuPDF==1.23.7
python-docx==0.8.11
spacy==3.7.2
python-dotenv==1.0.0
Pillow==10.0.1
fpdf2==2.7.8
werkzeug==2.3.7
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz